scraper/requirements.txt
# scraper/requirements.txt
requests
google-generativeai
python-dotenv
selenium
webdriver-manager
beautifulsoup4